Update executequery.md
IDictionary works fine, doesn't need to be a concrete Dictionary. This is good to know from an F# perspective as you can just do e.g.

```fsharp
let data = conn.ExecuteQuery("SELECT * FROM Capacity WHERE Col = @P1", dict [ "P1", box "Value" ])
```

`dict` is a built-in function in F# that takes in an IEnumerable of tuples and creates an IDictionary from it.